+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 0.2
+
+- Chg: Make `show` consistent for Page, View, PageSchema
+- Ref: Refactor `value` in `PropertyValue`
+- Ref: Reduce unnecessary funtionality of `Number`, use `value` instead
+
 ## Version 0.1, 2023-11-16
 
 - First official alpha release